fixed mse, cost classes
